final_pjt_back/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.3
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==4.0.1
Django==4.2.4
django-allauth==0.54.0
django-appconf==1.0.5
django-cors-headers==4.2.0
django-imagekit==4.1.0
djangorestframework==3.14.0
drf-spectacular==0.26.4
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.4
inflection==0.5.1
jiter==0.10.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
oauthlib==3.2.2
openai==1.82.0
pilkit==2.0
Pillow==8.3.2
pycparser==2.21
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.8.0
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.9.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.0.4
